requirements.txt
discord.py==2.0.0
aiocron==1.7
python-dotenv==0.19.2
pymongo==3.12.1
aiohttp==3.8.5
Flask==2.1.3
Werkzeug==2.0.3
random